\name{add.Point}
\alias{add.Point}
\title{Add a time point to a TextGrid}
\usage{
  add.Point(textgrid, tier, point, text)
}
\arguments{
  \item{textgrid}{The TextGrid object to modify.}

  \item{tier}{The name of the IntervalTier as a character
  vector.}

  \item{point}{The time point to be added.}

  \item{text}{The annotation for the point.}
}
\description{
  Adds a single time point to a specified IntervalTier
  within a TextGrid.
}
\author{
  John Pellman
}

