\name{annotations}
\alias{annotations}
\title{Extract Annotations}
\usage{
  annotations(textgrid, tier, tr)
}
\arguments{
  \item{textgrid}{The TextGrid object to modify.}

  \item{tier}{The name of the IntervalTier as a character
  vector.}

  \item{tr}{The time step for each index of the vector (in
  seconds).}
}
\value{
  A character vector containing the text annotations for
  each time step.
}
\description{
  Extract annotations into a vector
}
\author{
  John Pellman
}

