\name{modify.IntervalTier}
\alias{modify.IntervalTier}
\title{Change the name of an IntervalTier in a TextGrid}
\usage{
  remove.IntervalTier(textgrid, tier)
}
\arguments{
  \item{textgrid}{The TextGrid object to modify.}

  \item{oldname}{The current name of the IntervalTier.}

  \item{newname}{The new name for the IntervalTier.}
}
\description{
  Changes the name of an IntervalTier in a TextGrid.
}
\author{
  John Pellman
}

