\name{remove.IntervalTier}
\alias{remove.IntervalTier}
\title{Remove an IntervalTier from a TextGrid}
\usage{
  remove.IntervalTier(textgrid, tier)
}
\arguments{
  \item{textgrid}{The TextGrid object to modify.}

  \item{tier}{A character vector containing the name of the
  IntervalTier to be removed.}
}
\value{
  The TextGrid minus the specified IntervalTier to be
  removed.
}
\description{
  Removes an IntervalTier from a TextGrid.
}
\author{
  John Pellman
}

