\name{remove.Point}
\alias{remove.Point}
\title{Removes a time point from a TextGrid}
\usage{
  remove.Point(textgrid, tier, point, text)
}
\arguments{
  \item{textgrid}{The TextGrid object to modify.}

  \item{tier}{The name of the IntervalTier as a character
  vector.}

  \item{point}{The time point to be removed.}

  \item{text}{The annotation for the point.}
}
\description{
  Removes a single time point from a specified IntervalTier
  within a TextGrid.
}
\author{
  John Pellman
}

