\name{Rpraat-package}
\alias{Rpraat-package}
\alias{Rpraat}
\docType{package}
\title{Rpraat
}
\description{
Rpraat is a library that allows you to manipulate Praat TextGrid files in R.  Praat is an open source software package used extensively in phonetics research that allows you to annotate audio files of human speech.  The TextGrid format is a portable, self-contained representation of these annotations.
}
\details{
\tabular{ll}{
Package: \tab Rpraat\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-08-21\cr
License: \tab GPL v 3.0?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
John Pellman

Maintainer: John Pellman <pellman.john@gmail.com>
}

~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
