\name{textgridR-package}
\alias{textgridR-package}
\alias{textgridR}
\docType{package}
\title{textgridR
}
\description{
textgridR is a library that allows you to manipulate Praat TextGrid files in R.  Praat is an open source software package used extensively in phonetics research that allows you to annotate audio files of human speech.  The TextGrid format is a portable, self-contained representation of these annotations.
}
\details{
\tabular{ll}{
Package: \tab textgridR\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-08-21\cr
License: \tab GPL v 3.0?\cr
}
}
\author{
John Pellman

Maintainer: John Pellman <pellman.john@gmail.com>
}

\keyword{ package }
\seealso{
}
\examples{
}
