\name{write.TextGrid}
\alias{write.TextGrid}
\title{Export a TextGrid}
\usage{
  write.TextGrid(file)
}
\arguments{
  \item{file}{A file connection or a character vector
  containing the path to where the file should be saved.}
}
\value{
  A logical indicating whether or not the write was
  successful.
}
\description{
  Exports a TextGrid file to a specified file path.
}
\author{
  John Pellman
}

