\name{add.Interval}
\alias{add.Interval}
\title{Add an Interval to a TextGrid}
\usage{
  add.Interval(tg, it)
}
\arguments{
  \item{tg}{The TextGrid object to modify.}

  \item{it}{The name of the IntervalTier as a character
  vector.}
}
\description{
  Adds an Interval to a specified IntervalTier within a
  TextGrid.
}
\author{
  John Pellman
}

